S_KaladaiYT/Level2_QuizPrep/Session_06_ModelSetting/pyproject.toml
[project]
name = "session-06-modelsetting"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openai-agents==0.2.11",
    "pydantic>=2.11.10",
    "python-dotenv>=1.1.1",
    "rich>=14.1.0",
]
